third_party/WebKit/LayoutTests/http/tests/inspector/network/network-blocked-reason-expected.txt

CONSOLE ERROR: Refused to load the script 'https://www.example.com/csp.js' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

Tests that blocked reason is recognized correctly.


Running: testCSP

Request: csp.js
BlockedReason: csp

Running: testDevTools

Request: silent_script.js
BlockedReason: inspector

Request: silent_script.js
BlockedReason: undefined
